# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Exam Preparation System (EPS) API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      parameters:
        - in: query
          name: role
          schema:
            $ref: "#/components/schemas/Role"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createUser
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Email already in use
  /admin/users/{id}:
    patch:
      operationId: updateUser
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteUser
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
        "400":
          description: Bad request
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/deletion-requests:
    get:
      operationId: listDeletionRequests
      tags: [admin]
      summary: All account deletion requests (newest first)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeletionRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [admin]
      summary: System overview totals (users, courses, topics, questions, exams)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
  responses:
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
    Role:
      type: string
      enum: [student, lecturer, admin]
...
    User:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        accountStatus: { type: string }
        profileImageUrl: { type: string, nullable: true }
        programId: { type: integer, nullable: true, description: "Student's program (null for admins/lecturers)" }
        programName: { type: string, nullable: true }
        programCode: { type: string, nullable: true }
        programIds:
          type: array
          description: "Programs a lecturer teaches in (only populated for lecturers)"
          items: { type: integer }
      required: [id, fullName, email, role, accountStatus]
...
    CreateUserRequest:
      type: object
      properties:
        fullName: { type: string }
        email: { type: string }
        password: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        accountStatus: { type: string }
        programId:
          type: integer
          nullable: true
          description: "For role=student: the program they study in"
        programIds:
          type: array
          description: "For role=lecturer: programs they teach in"
          items: { type: integer }
      required: [fullName, email, password, role]
    UpdateUserRequest:
      type: object
      properties:
        fullName: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        accountStatus: { type: string }
...
    DeletionRequest:
      type: object
      properties:
        id: { type: integer }
        formerUserId: { type: integer }
        formerEmail: { type: string }
        formerFullName: { type: string }
        formerRole: { $ref: "#/components/schemas/Role" }
        reason: { type: string }
        deletedAt: { type: string, format: date-time }
      required: [id, formerUserId, formerEmail, formerFullName, formerRole, reason, deletedAt]
...
    AdminOverview:
      type: object
      properties:
        totals:
          type: object
          properties:
            users: { type: integer }
            students: { type: integer }
            lecturers: { type: integer }
            admins: { type: integer }
            courses: { type: integer }
            topics: { type: integer }
            questions: { type: integer }
            approvedQuestions: { type: integer }
            archivedQuestions: { type: integer }
            exams: { type: integer }
            submittedExams: { type: integer }
          required:
            [users, students, lecturers, admins, courses, topics, questions, approvedQuestions, archivedQuestions, exams, submittedExams]
      required: [totals]