# lib/api-zod/src/generated/types/createUserRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Exam Preparation System (EPS) API
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface CreateUserRequest {
  fullName: string;
  email: string;
  password: string;
  role: Role;
  accountStatus?: string;
  /** For role=student: the program they study in */
  programId?: number | null;
  /** For role=lecturer: programs they teach in */
  programIds?: number[];
}