# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Exam Preparation System (EPS) API
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface User {
  id: number;
  fullName: string;
  email: string;
  role: Role;
  accountStatus: string;
  profileImageUrl?: string | null;
  /** Student's program (null for admins/lecturers) */
  programId?: number | null;
  programName?: string | null;
  programCode?: string | null;
  /** Programs a lecturer teaches in (only populated for lecturers) */
  programIds?: number[];
}